Create Probing Agent
 Probing script probes an SNMP agent and finds the rate of change for several counters between successive probes/ samples. The rate calculated for each counter/OID will be displayed on the console, one line for each calculated rate, the output format will be described in detail in 'output format'.

Output format
The output from the script  as follows:
Sample time | OID1 | OID2 | .... | OIDn

 Sample time: Timestamp of the last sample, in UNIX time (seconds). 
OID*: Rate of OID* between the last two successful samples

As an example:
1303567100  | 1234 | 456 | 1567 | 78903 
1303567101  | 2345 | 234 | 2356 | 67892 
1303567102  | 3456 | 345 | 894 | 23467 

The script will be invoked as follows:
prober <Agent IP:port:community> <sample frequency> <samples> <OID1> <OID2> …….. <OIDn>
where,
IP, port and community are agent details,
OIDn are the OIDs to be probed (they are absolute, cf. IF-MIB::ifInOctets.2 for interface 2, or 1.3.6.1.2.1.2.2.1.10.2 [1]) 
Sample frequency  (Fs) is the sampling frequency expressed in Hz, you should handle between 10 and 0.1 Hz. 
Samples (N) is the number of successful samples the solution should do before terminating, hence the value should be greater or equal to 2. If the value is -1 that means run forever (until CTRL-C is pressed, or the app is terminated in some way).