Create Traphandling
The traphandling script receives the trap signals from different servers with specified OID and inserts it into SqLite3 database. Then it sees the status in the 2nd row of the table and generates a trap signal (if status=3 ) (if status=2  & previous status!= 3) and sends to a server which is present in another table Krishna in the same database.  

It listens in a different port.

This script is linked with Getstatus, setTrapR, getTrapR script.