Create Getstatus
This script has no input, but the the output is either FALSE (if no devices have been logged) or a list of devices in the the following format of: 

<device name1> | <status> | <report time of status> | <oldstatus (if applicable)> | <old report time (if applicable)>
<device name2> | <status> | <report time of status> | <oldstatus (if applicable)> | <old report time (if applicable)>

....

<device nameK> | <status> | <report time of status> | <oldstatus (if applicable)> | <old report time (if applicable)>

 

where 

the device name is the name provided in the trap

status is the numerical status provided in the trap

report time of status is the time the trap containing the status was received by the system. 

old status is the previous status if the device had any.

old report time is the time when the old data was reported.